More GUI client controller work
